Factor out common code in NewmanPenrose and Gauge transform tests
