+++ tests/Unit/Evolution/Systems/Cce/Test_GaugeTransformBoundaryData.cpp
@@ -3,76 +3,23 @@
 
 #include "Framework/TestingFramework.hpp"
 
+#include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataBox/Prefixes.hpp"
-#include "DataStructures/DataBox/TagName.hpp"
-#include "DataStructures/VariablesTag.hpp"
-#include "Evolution/Systems/Cce/BoundaryData.hpp"
 #include "Evolution/Systems/Cce/GaugeTransformBoundaryData.hpp"
-#include "Evolution/Systems/Cce/Initialize/InitializeJ.hpp"
-#include "Evolution/Systems/Cce/OptionTags.hpp"
 #include "Evolution/Systems/Cce/PreSwshDerivatives.hpp"
 #include "Evolution/Systems/Cce/Tags.hpp"
 #include "Framework/TestHelpers.hpp"
-#include "Helpers/DataStructures/MakeWithRandomValues.hpp"
-#include "Helpers/Evolution/Systems/Cce/BoundaryTestHelpers.hpp"
+#include "Helpers/Evolution/Systems/Cce/VolumeTestHelpers.hpp"
 #include "NumericalAlgorithms/RootFinding/TOMS748.hpp"
-#include "NumericalAlgorithms/Spectral/Basis.hpp"
-#include "NumericalAlgorithms/Spectral/Quadrature.hpp"
-#include "NumericalAlgorithms/Spectral/Spectral.hpp"
-#include "PointwiseFunctions/AnalyticSolutions/GeneralRelativity/KerrSchild.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/VectorAlgebra.hpp"
 
 namespace Cce {
 namespace {
-
-struct InverseCubicEvolutionGauge {
-  using boundary_tags =
-      tmpl::list<Tags::EvolutionGaugeBoundaryValue<Tags::BondiJ>,
-                 Tags::EvolutionGaugeBoundaryValue<Tags::Dr<Tags::BondiJ>>,
-                 Tags::EvolutionGaugeBoundaryValue<Tags::BondiR>>;
-
-  using mutate_tags = tmpl::list<Tags::BondiJ, Tags::CauchyCartesianCoords,
-                                 Tags::CauchyAngularCoords>;
-  using argument_tags =
-      tmpl::append<boundary_tags,
-                   tmpl::list<Tags::LMax, Tags::NumberOfRadialPoints>>;
-
-  InverseCubicEvolutionGauge() = default;
-
-  void operator()(
-      const gsl::not_null<Scalar<SpinWeighted<ComplexDataVector, 2>>*> j,
-      const gsl::not_null<tnsr::i<DataVector, 3>*> /*cartesian_x_of_x_tilde*/,
-      const gsl::not_null<
-          tnsr::i<DataVector, 2, ::Frame::Spherical<::Frame::Inertial>>*>
-      /*angular_cauchy_coordinates*/,
-      const Scalar<SpinWeighted<ComplexDataVector, 2>>& boundary_j,
-      const Scalar<SpinWeighted<ComplexDataVector, 2>>& boundary_dr_j,
-      const Scalar<SpinWeighted<ComplexDataVector, 0>>& r,
-      const size_t /*l_max*/, const size_t number_of_radial_points) const {
-    const DataVector one_minus_y_collocation =
-        1.0 - Spectral::collocation_points<Spectral::Basis::Legendre,
-                                           Spectral::Quadrature::GaussLobatto>(
-                  number_of_radial_points);
-    for (size_t i = 0; i < number_of_radial_points; i++) {
-      ComplexDataVector angular_view_j{
-          get(*j).data().data() + get(boundary_j).size() * i,
-          get(boundary_j).size()};
-      // auto is acceptable here as these two values are only used once in the
-      // below computation. `auto` causes an expression template to be
-      // generated, rather than allocating.
-      const auto one_minus_y_coefficient =
-          0.25 * (3.0 * get(boundary_j).data() +
-                  get(r).data() * get(boundary_dr_j).data());
-      const auto one_minus_y_cubed_coefficient =
-          -0.0625 *
-          (get(boundary_j).data() + get(r).data() * get(boundary_dr_j).data());
-      angular_view_j =
-          one_minus_y_collocation[i] * one_minus_y_coefficient +
-          pow<3>(one_minus_y_collocation[i]) * one_minus_y_cubed_coefficient;
-    }
-  }
-};
+using coordinate_variables_tag = TestHelpers::coordinate_variables_tag;
+using spin_weighted_variables_tag = TestHelpers::spin_weighted_variables_tag;
+using volume_spin_weighted_variables_tag =
+    TestHelpers::volume_spin_weighted_variables_tag;
 
 // These gauge transforms are extremely hard to validate outside of a true
 // evolution system. Here we settle for verifying that for an
@@ -88,174 +35,8 @@ void test_gauge_transforms_via_inverse_coordinate_map(
   const size_t number_of_angular_grid_points =
       Spectral::Swsh::number_of_swsh_collocation_points(l_max);
 
-  using real_boundary_tags =
-      tmpl::list<Tags::CauchyAngularCoords, Tags::CauchyCartesianCoords,
-                 Tags::PartiallyFlatAngularCoords,
-                 Tags::PartiallyFlatCartesianCoords,
-                 ::Tags::dt<Tags::CauchyCartesianCoords>,
-                 ::Tags::dt<Tags::PartiallyFlatCartesianCoords>>;
-  using spin_weighted_boundary_tags = tmpl::flatten<tmpl::list<
-      tmpl::list<Tags::PartiallyFlatGaugeC, Tags::PartiallyFlatGaugeD,
-                 Tags::PartiallyFlatGaugeOmega, Tags::CauchyGaugeC,
-                 Tags::CauchyGaugeD, Tags::CauchyGaugeOmega,
-                 Tags::Du<Tags::PartiallyFlatGaugeOmega>,
-                 Spectral::Swsh::Tags::Derivative<Tags::PartiallyFlatGaugeOmega,
-                                                  Spectral::Swsh::Tags::Eth>,
-                 Spectral::Swsh::Tags::Derivative<Tags::CauchyGaugeOmega,
-                                                  Spectral::Swsh::Tags::Eth>,
-                 Tags::BondiUAtScri>,
-      Tags::characteristic_worldtube_boundary_tags<Tags::BoundaryValue>,
-      Tags::characteristic_worldtube_boundary_tags<
-          Tags::EvolutionGaugeBoundaryValue>>>;
-  using coordinate_variables_tag = ::Tags::Variables<real_boundary_tags>;
-  using spin_weighted_variables_tag =
-      ::Tags::Variables<spin_weighted_boundary_tags>;
-  using volume_spin_weighted_variables_tag = ::Tags::Variables<
-      tmpl::list<Tags::BondiJ, Tags::Dy<Tags::BondiJ>, Tags::BondiU>>;
-
-  auto forward_transform_box = db::create<db::AddSimpleTags<
-      coordinate_variables_tag, spin_weighted_variables_tag,
-      volume_spin_weighted_variables_tag, Tags::LMax,
-      Tags::NumberOfRadialPoints,
-      Spectral::Swsh::Tags::SwshInterpolator<Tags::CauchyAngularCoords>,
-      Spectral::Swsh::Tags::SwshInterpolator<
-          Tags::PartiallyFlatAngularCoords>>>(
-      typename coordinate_variables_tag::type{number_of_angular_grid_points},
-      typename spin_weighted_variables_tag::type{number_of_angular_grid_points},
-      typename volume_spin_weighted_variables_tag::type{
-          number_of_angular_grid_points * number_of_radial_grid_points},
-      l_max, number_of_radial_grid_points, Spectral::Swsh::SwshInterpolator{},
-      Spectral::Swsh::SwshInterpolator{});
-
-  // create analytic data for the forward transform
-  UniformCustomDistribution<double> value_dist{0.1, 0.5};
-  // first prepare the input for the modal version
-  const double mass = value_dist(*gen);
-  const std::array<double, 3> spin{
-      {value_dist(*gen), value_dist(*gen), value_dist(*gen)}};
-  const std::array<double, 3> center{
-      {value_dist(*gen), value_dist(*gen), value_dist(*gen)}};
-  const gr::Solutions::KerrSchild solution{mass, spin, center};
-
-  const double extraction_radius = 100.0;
-
-  // acceptable parameters for the fake sinusoid variation in the input
-  // parameters
-  const double frequency = 0.1 * value_dist(*gen);
-  const double amplitude = 0.1 * value_dist(*gen);
-  const double target_time = 50.0 * value_dist(*gen);
-
-  const size_t libsharp_size =
-      Spectral::Swsh::size_of_libsharp_coefficient_vector(l_max);
-  tnsr::ii<ComplexModalVector, 3> spatial_metric_coefficients{libsharp_size};
-  tnsr::ii<ComplexModalVector, 3> dt_spatial_metric_coefficients{libsharp_size};
-  tnsr::ii<ComplexModalVector, 3> dr_spatial_metric_coefficients{libsharp_size};
-  tnsr::I<ComplexModalVector, 3> shift_coefficients{libsharp_size};
-  tnsr::I<ComplexModalVector, 3> dt_shift_coefficients{libsharp_size};
-  tnsr::I<ComplexModalVector, 3> dr_shift_coefficients{libsharp_size};
-  Scalar<ComplexModalVector> lapse_coefficients{libsharp_size};
-  Scalar<ComplexModalVector> dt_lapse_coefficients{libsharp_size};
-  Scalar<ComplexModalVector> dr_lapse_coefficients{libsharp_size};
-  TestHelpers::create_fake_time_varying_data(
-      make_not_null(&spatial_metric_coefficients),
-      make_not_null(&dt_spatial_metric_coefficients),
-      make_not_null(&dr_spatial_metric_coefficients),
-      make_not_null(&shift_coefficients), make_not_null(&dt_shift_coefficients),
-      make_not_null(&dr_shift_coefficients), make_not_null(&lapse_coefficients),
-      make_not_null(&dt_lapse_coefficients),
-      make_not_null(&dr_lapse_coefficients), solution, extraction_radius,
-      amplitude, frequency, target_time, l_max, false);
-
-  db::mutate<spin_weighted_variables_tag>(
-      [&spatial_metric_coefficients, &dt_spatial_metric_coefficients,
-       &dr_spatial_metric_coefficients, &shift_coefficients,
-       &dt_shift_coefficients, &dr_shift_coefficients, &lapse_coefficients,
-       &dt_lapse_coefficients, &dr_lapse_coefficients, &extraction_radius](
-          const gsl::not_null<Variables<spin_weighted_boundary_tags>*>
-              spin_weighted_boundary_variables) {
-        create_bondi_boundary_data(
-            spin_weighted_boundary_variables, spatial_metric_coefficients,
-            dt_spatial_metric_coefficients, dr_spatial_metric_coefficients,
-            shift_coefficients, dt_shift_coefficients, dr_shift_coefficients,
-            lapse_coefficients, dt_lapse_coefficients, dr_lapse_coefficients,
-            extraction_radius, l_max);
-      },
-      make_not_null(&forward_transform_box));
-
-  // construct the coordinate transform quantities
-  const double variation_amplitude = value_dist(*gen);
-  const double variation_amplitude_inertial = value_dist(*gen);
-  db::mutate<Tags::CauchyCartesianCoords>(
-      [&l_max,
-       &variation_amplitude](const gsl::not_null<tnsr::i<DataVector, 3>*>
-                                 cauchy_cartesian_coordinates) {
-        tnsr::i<DataVector, 2> cauchy_angular_coordinates{
-            get<0>(*cauchy_cartesian_coordinates).size()};
-        // There is a bug in Clang 10.0.0 that gives a nonsensical
-        // error message for the following call to
-        // cached_collocation_metadata unless l_max is captured in
-        // this lambda. The capture should not be necessary because l_max is a
-        // const integer type that is initialized by a constant expression. Note
-        // that l_max need not be declared constexpr for its value to be
-        // retrieved inside a lambda without capturing it. This
-        // line silences the warning that says capturing l_max is not necessary.
-        (void)l_max;
-        const auto& collocation = Spectral::Swsh::cached_collocation_metadata<
-            Spectral::Swsh::ComplexRepresentation::Interleaved>(l_max);
-        for (const auto collocation_point : collocation) {
-          get<1>(cauchy_angular_coordinates)[collocation_point.offset] =
-              collocation_point.phi + 1.0e-2 * variation_amplitude *
-                                          cos(collocation_point.phi) *
-                                          sin(collocation_point.theta);
-          get<0>(cauchy_angular_coordinates)[collocation_point.offset] =
-              collocation_point.theta;
-        }
-        get<0>(*cauchy_cartesian_coordinates) =
-            sin(get<0>(cauchy_angular_coordinates)) *
-            cos(get<1>(cauchy_angular_coordinates));
-        get<1>(*cauchy_cartesian_coordinates) =
-            sin(get<0>(cauchy_angular_coordinates)) *
-            sin(get<1>(cauchy_angular_coordinates));
-        get<2>(*cauchy_cartesian_coordinates) =
-            cos(get<0>(cauchy_angular_coordinates));
-      },
-      make_not_null(&forward_transform_box));
-
-  db::mutate<Tags::PartiallyFlatCartesianCoords>(
-      [&l_max, &variation_amplitude_inertial](
-          const gsl::not_null<tnsr::i<DataVector, 3>*>
-              inertial_cartesian_coordinates) {
-        tnsr::i<DataVector, 2> inertial_angular_coordinates{
-            get<0>(*inertial_cartesian_coordinates).size()};
-        // There is a bug in Clang 10.0.0 that gives a nonsensical
-        // error message for the following call to
-        // cached_collocation_metadata unless l_max is captured in
-        // this lambda. The capture should not be necessary because l_max is a
-        // const integer type that is initialized by a constant expression. Note
-        // that l_max need not be declared constexpr for its value to be
-        // retrieved inside a lambda without capturing it. This
-        // line silences the warning that says capturing l_max is not necessary.
-        (void)l_max;
-        const auto& collocation = Spectral::Swsh::cached_collocation_metadata<
-            Spectral::Swsh::ComplexRepresentation::Interleaved>(l_max);
-        for (const auto& collocation_point : collocation) {
-          get<1>(inertial_angular_coordinates)[collocation_point.offset] =
-              collocation_point.phi + 1.0e-2 * variation_amplitude_inertial *
-                                          cos(collocation_point.phi) *
-                                          sin(collocation_point.theta);
-          get<0>(inertial_angular_coordinates)[collocation_point.offset] =
-              collocation_point.theta;
-        }
-        get<0>(*inertial_cartesian_coordinates) =
-            sin(get<0>(inertial_angular_coordinates)) *
-            cos(get<1>(inertial_angular_coordinates));
-        get<1>(*inertial_cartesian_coordinates) =
-            sin(get<0>(inertial_angular_coordinates)) *
-            sin(get<1>(inertial_angular_coordinates));
-        get<2>(*inertial_cartesian_coordinates) =
-            cos(get<0>(inertial_angular_coordinates));
-      },
-      make_not_null(&forward_transform_box));
+  auto forward_transform_box = Cce::TestHelpers::create_cce_volume_box(
+      gen, l_max, number_of_radial_grid_points, false);
 
   auto inverse_transform_box = db::create<db::AddSimpleTags<
       coordinate_variables_tag, spin_weighted_variables_tag,
@@ -270,6 +51,12 @@ void test_gauge_transforms_via_inverse_coordinate_map(
           number_of_angular_grid_points * number_of_radial_grid_points},
       l_max, number_of_radial_grid_points, Spectral::Swsh::SwshInterpolator{},
       Spectral::Swsh::SwshInterpolator{});
+
+  const double variation_amplitude =
+      db::get<Cce::TestHelpers::VariationAmplitude>(forward_transform_box);
+  const double variation_amplitude_inertial =
+      db::get<Cce::TestHelpers::InertialVariationAmplitude>(
+          forward_transform_box);
 
   db::mutate<Tags::CauchyCartesianCoords>(
       [&l_max,
@@ -626,6 +413,8 @@ void test_gauge_transforms_via_inverse_coordinate_map(
         check_gauge_adjustment_against_inverse);
   }
 
+  using InverseCubicEvolutionGauge =
+      Cce::TestHelpers::InverseCubicEvolutionGauge;
   db::mutate_apply<InverseCubicEvolutionGauge::mutate_tags,
                    InverseCubicEvolutionGauge::argument_tags>(
       InverseCubicEvolutionGauge{}, make_not_null(&forward_transform_box));
